refactor(p2p): event_handler.go
